Fix checking :custom_slug existence to use column_exists?

+++ pages/db/migrate/20140105190324_add_custom_slug_to_refinery_pages.rb
@@ -1,20 +1,15 @@
 class AddCustomSlugToRefineryPages < ActiveRecord::Migration[4.2]
   def up
-    if page_column_names.exclude?('custom_slug')
-      add_column :refinery_pages, :custom_slug, :string
-    end
+    add_column :refinery_pages, :custom_slug, :string unless custom_slug_exists?
   end
 
   def down
-    if page_column_names.include?('custom_slug')
-      remove_column :refinery_pages, :custom_slug
-    end
+    remove_column :refinery_pages, :custom_slug if custom_slug_exists?
   end
 
   private
-  def page_column_names
-    return [] unless defined?(::Refinery::Page)
 
-    Refinery::Page.column_names.map(&:to_s)
+  def custom_slug_exists?
+    column_exists?(:refinery_pages, :custom_slug)
   end
 end
